+++ src/main.ts
@@ -38,6 +38,7 @@ export * from './components/Modal';
 export * from './components/Overlay';
 export * from './components/Panel';
 export * from './components/Pagination';
+export * from './components/Skeleton';
 export * from './components/Steps';
 export * from './components/Table';
 export * from './components/Tile';